webapp/requirements.txt
blinker==1.7.0
Flask==3.0.0
importlib-metadata==6.8.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
osrf-pycommon==2.1.3
Pillow==10.1.0
pypng==0.20220715.0
PyQRCode==1.2.1
Werkzeug==3.0.1
zipp==3.17.0
pyzbar==0.1.9